[DOC] Drink.js
